Huge update adressing the moving or rotating the object from edit mode
Added a tracking of active object location and rotation to know if it's origin was altered from the object properties panel or  with an addon. While in 'Edit' mode the  'ghost_to_cursor' operator has to call an 'undo' to be properly registred by Blender's undo system, and that interfered with addons that temporarily switched the modes to 'Object' and back, or somehow manipulate with object's origin in other way. Tracking helps to handle this situation and seems to be working.

Added 'poll' variable to prevent any possible issues caused by 'Redo Last' (default: Shift+R) called for 'ghost_to_cursor' operator.

Overall code cleanup